+++ Script/Wobbler.cs
@@ -16,26 +16,26 @@ namespace Nebulator.Script
         Random _rand = new Random();
 
         /// <summary>Minimum range for randomizing - 3 sigma.</summary>
-        public int RangeLow { get; set; } = 0;
+        public double RangeLow { get; set; } = 0;
 
         /// <summary>Maximum range for randomizing - 3 sigma.</summary>
-        public int RangeHigh { get; set; } = 0;
+        public double RangeHigh { get; set; } = 0;
 
         /// <summary>
         /// Return next from standard distribution.
         /// </summary>
         /// <param name="val">Center distribution around this.</param>
         /// <returns>Randomized value or val if ranges are 0 (default).</returns>
-        public int Next(int val)
+        public double Next(double val)
         {
-            int newVal = val; // default
+            double newVal = val; // default
 
             if (RangeLow != 0 || RangeHigh != 0)
             {
-                int max = val + RangeHigh;
-                int min = val + RangeLow;
-                int mean = min + (max - min) / 2;
-                int sigma = (max - min) / 3; // 3 sd
+                double max = val + RangeHigh;
+                double min = val + RangeLow;
+                double mean = min + (max - min) / 2;
+                double sigma = (max - min) / 3; // 3 sd
                 newVal = (int)Utils.NextGaussian(_rand, mean, sigma);
             }
             return newVal;
@@ -48,7 +48,7 @@ namespace Nebulator.Script
         /// <returns>Randomized value or val if Min/Max are 0.</returns>
         public Time Next(Time val)
         {
-            int newVal = Next(val.TotalTocks);
+            int newVal = (int)Next(val.TotalTocks);
             return new Time(newVal);
         }
     }
